refactor src/components/Notification/messages.tsx into messages.tsx and helpers.tsx src/utils
