+++ src/utils/helpers.ts
@@ -2,6 +2,23 @@ import { IMapPin } from 'src/models/maps.models'
 import { IUser } from 'src/models/user.models'
 import { DBDoc, IModerable } from 'src/models/common.models'
 import { toJS, isObservableObject } from 'mobx'
+import MESSAGES from './messages'
+
+/**
+ * Conversion for default error messages.
+ * @param systemMessage - the message text for lookup in the table.
+ * This can either be a status code or full message (depending on how saved above)
+ */
+export const getFriendlyMessage = (systemMessage: string) => {
+  const friendlyMessage = MESSAGES[systemMessage.toLowerCase()]
+  if (!friendlyMessage) {
+    console.log(
+      `%c No friendly message for [${systemMessage}] \n Maybe you should add one?`,
+      'background: #222; color: #bada55',
+    )
+  }
+  return friendlyMessage ? friendlyMessage : systemMessage
+}
 
 // remove special characters from string, also replacing spaces with dashes
 export const stripSpecialCharacters = (text: string) => {
